packaging: add empty directories
Empty directories are not tracked in git, but the build relies on them. Add
a placeholder file to each in order to have them added to git.

This task has been done by extracting the tarball of the upstream, then running:

find ./ -type d -empty > /tmp/empty_dirs

Which report the empty directories.

Then, by a for loop, the directories and the placeholders were created:

for i in $(cat /tmp/empty_dirs) ; do
    echo $i;
    mkdir -p $i;
    echo "This file is here so git creates the parent directory." > $i/.empty_directory;
done